1-Models/1.1-LLMs/1-llm-demo.py
from langchain_openai import OpenAI

from dotenv import load_dotenv
import os

# Load .env from parent directory
load_dotenv(dotenv_path=os.path.join(os.path.dirname(__file__), '..', '.env'))

llm = OpenAI(model='gpt-3.5-turbo-instruct')

result = llm.invoke("What is the capital of India?")

print(result)